Move merge_bams to bamProcessing
